# packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Getter API",
    "version": "0.0.0",
    "description": "Generated from the TypeBox contracts in packages/contracts by packages/contracts/scripts/generate.ts. Do not edit by hand. JSON snake_case, ISO-8601 UTC timestamps, UUID identifiers. The API derives the workspace from the authenticated principal and never trusts a client workspace_id."
  },
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jg_session",
        "description": "HttpOnly session cookie issued by POST /auth/login."
      },
      "deviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped device token for the extension or local runner; expires after 30 days."
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Operator worker credential for the internal task protocol only."
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "anyOf": [
                  {
                    "const": "VALIDATION_ERROR",
                    "type": "string"
                  },
                  {
                    "const": "MALFORMED_REQUEST",
                    "type": "string"
                  },
                  {
                    "const": "UNAUTHENTICATED",
                    "type": "string"
                  },
                  {
                    "const": "FORBIDDEN",
                    "type": "string"
                  },
                  {
                    "const": "NOT_FOUND",
                    "type": "string"
                  },
                  {
                    "const": "CONFLICT",
                    "type": "string"
                  },
                  {
                    "const": "STALE_REVISION",
                    "type": "string"
                  },
                  {
                    "const": "IDEMPOTENCY_MISMATCH",
                    "type": "string"
                  },
                  {
                    "const": "PAYLOAD_TOO_LARGE",
                    "type": "string"
                  },
                  {
                    "const": "UNPROCESSABLE",
                    "type": "string"
                  },
                  {
                    "const": "QUOTA_EXCEEDED",
                    "type": "string"
                  },
                  {
                    "const": "SETUP_CLOSED",
                    "type": "string"
                  },
                  {
                    "const": "BUDGET_EXHAUSTED",
                    "type": "string"
                  },
                  {
                    "const": "PROVIDER_UNAVAILABLE",
                    "type": "string"
                  },
                  {
                    "const": "INTERNAL_ERROR",
                    "type": "string"
                  }
                ]
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "patternProperties": {
                  "^(.*)$": {
                    "type": "string"
                  }
                }
              },
              "request_id": {
                "format": "uuid",
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProviderLimits": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "context_limit",
          "output_token_limit",
          "temperature",
          "timeout_seconds",
          "daily_token_budget",
          "daily_cost_budget"
        ],
        "properties": {
          "context_limit": {
            "minimum": 512,
            "maximum": 2000000,
            "type": "integer"
          },
          "output_token_limit": {
            "minimum": 64,
            "maximum": 200000,
            "type": "integer"
          },
          "temperature": {
            "minimum": 0,
            "maximum": 2,
            "type": "number"
          },
          "timeout_seconds": {
            "minimum": 5,
            "maximum": 600,
            "type": "integer"
          },
          "daily_token_budget": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "daily_cost_budget": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "ProviderSettingsPutRequest": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "model",
          "base_url",
          "limits",
          "rate_card"
        ],
        "properties": {
          "provider": {
            "anyOf": [
              {
                "const": "none",
                "type": "string"
              },
              {
                "const": "fake",
                "type": "string"
              },
              {
                "const": "ollama",
                "type": "string"
              },
              {
                "const": "openai_compatible",
                "type": "string"
              }
            ]
          },
          "model": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "base_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "api_key": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "limits": {
            "$ref": "#/components/schemas/ProviderLimits"
          },
          "rate_card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RateCard"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "ProviderSettingsView": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "model",
          "base_url",
          "api_key_set",
          "api_key_masked",
          "limits",
          "rate_card",
          "sends_data_externally",
          "updated_at"
        ],
        "properties": {
          "provider": {
            "anyOf": [
              {
                "const": "none",
                "type": "string"
              },
              {
                "const": "fake",
                "type": "string"
              },
              {
                "const": "ollama",
                "type": "string"
              },
              {
                "const": "openai_compatible",
                "type": "string"
              }
            ]
          },
          "model": {
            "type": "string"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "api_key_set": {
            "type": "boolean"
          },
          "api_key_masked": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "limits": {
            "$ref": "#/components/schemas/ProviderLimits"
          },
          "rate_card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RateCard"
              },
              {
                "type": "null"
              }
            ]
          },
          "sends_data_externally": {
            "type": "boolean"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "ProviderTestResult": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reachable",
          "structured_output_supported",
          "model_available",
          "latency_ms",
          "detail"
        ],
        "properties": {
          "reachable": {
            "type": "boolean"
          },
          "structured_output_supported": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ]
          },
          "model_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ]
          },
          "latency_ms": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "maxLength": 1000,
            "type": "string"
          }
        }
      },
      "RateCard": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "currency",
          "input_cost_per_million",
          "output_cost_per_million"
        ],
        "properties": {
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "type": "string"
          },
          "input_cost_per_million": {
            "minimum": 0,
            "type": "number"
          },
          "output_cost_per_million": {
            "minimum": 0,
            "type": "number"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/settings/providers": {
      "get": {
        "operationId": "getProviderSettings",
        "summary": "Provider configuration with masked readiness only.",
        "tags": [
          "settings"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provider configuration with masked readiness only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderSettingsView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "putProviderSettings",
        "summary": "Store provider configuration; secrets are write-only.",
        "tags": [
          "settings"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Store provider configuration; secrets are write-only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderSettingsView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Anti-CSRF token for cookie-authenticated mutations; origin is verified too.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSettingsPutRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/providers/test": {
      "post": {
        "operationId": "testProviderSettings",
        "summary": "Probe the configured provider only; not a URL fetcher.",
        "tags": [
          "settings"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Probe the configured provider only; not a URL fetcher.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderTestResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Anti-CSRF token for cookie-authenticated mutations; origin is verified too.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {}
              }
            }
          }
        }
      }
    },
...
  }
}